exception handler

+++ services/file_service.py
@@ -5,7 +5,6 @@ from utils import utility_functions
 from models import FileUploadRequest, FileDownloadResponse, FileOperationResponse
 from models import FileNotFoundError, FileOperationError
 
-
 logger = utility_functions.create_logger()
 config = utility_functions.load_config()
 
@@ -24,8 +23,11 @@ class FileService:
 
     async def download_file(self, filename: str) -> Optional[FileDownloadResponse]:
         """Download the content of a file"""
+        file_path = os.path.join(self.base_dir, filename)
+        if not os.path.exists(file_path):
+            logger.error(f"File '{filename}' not found.")
+            raise FileNotFoundError(filename)
         try:
-            file_path = os.path.join(self.base_dir, filename)
             async with aiofiles.open(file_path, 'rb') as source_file:
                 async with aiofiles.open(config['DOWNLOAD_PATH'], 'wb') as dest_file:
                     chunk = await source_file.read(config['KILOBYTE'])
@@ -33,20 +35,15 @@ class FileService:
                         await dest_file.write(chunk)
                         chunk = await source_file.read(config['KILOBYTE'])
             return FileDownloadResponse(filename=filename, message="downloaded")
-        except FileNotFoundError as e:
-            logger.error(f"The file was not found: {e}")
-            return None
-        except OSError as e:
-            logger.error(f"The file was not found: {e}")
 
-            return None
+        except OSError as e:
+            logger.error(f"Error downloading file '{filename}': {e}")
+            raise FileOperationError(f"Error downloading file '{filename}'.")
 
     async def upload_file(self, request: FileUploadRequest) -> Optional[FileOperationResponse]:
-
         """Save content to a file"""
-
+        file_path = os.path.join(self.base_dir, request.filename)
         try:
-            file_path = os.path.join(self.base_dir, request.filename)
             async with aiofiles.open(request.content, 'rb') as source_file:
                 async with aiofiles.open(file_path, 'wb') as dest_file:
                     chunk = await source_file.read(config['KILOBYTE'])
@@ -56,17 +53,18 @@ class FileService:
             return FileOperationResponse(status="success", filename=request.filename, path=file_path)
         except OSError as e:
             logger.error(f"Error saving file {request.filename}: {e}")
-            return FileOperationResponse(status="error", detail=str(e))
+            raise FileOperationError(f"Error saving file '{request.filename}'.")
 
     def delete_file(self, filename: str) -> FileOperationResponse:
         """Delete a file"""
+        file_path = os.path.join(self.base_dir, filename)
+        if not os.path.exists(file_path):
+            logger.error(f"File '{filename}' not found.")
+            raise FileNotFoundError(filename)
         try:
-            file_path = os.path.join(self.base_dir, filename)
             os.remove(file_path)
             return FileOperationResponse(status="success", filename=filename)
-        except FileNotFoundError:
-            logger.error(f"File not found: {filename}")
-            return FileOperationResponse(status="error", detail="File not found")
+
         except OSError as e:
             logger.error(f"Error deleting file {filename}: {e}")
-            return FileOperationResponse(status="error", detail=str(e))
+            raise FileOperationError(f"Error deleting file '{filename}'.")
